refactor(ui): move works count label out of domain
- Place formatWorksCountLabel under lib/ui/formatters (presentation copy, not domain).
- Relocate unit tests to match.
- Add widget tests for PlaylistDetailsHeader singular/plural work labels.

Addresses PR #313 review feedback.

Made-with: Cursor
